requirements.txt
flask
mysql-connector-python
openai
python-dotenv
requests